staging: android: lowmemorykiller: Management of not killable processes
This feature allows the user to define processes and system processes
white-lists filled with the preferred process names and to not be killed
by lowmemorykiller, unless it is absolutely necessary.

This feature exposes four new lowmemorykiller parameters:
- donotkill_proc, is the flag to enable / disable this feature for the
processes chosen by the user
- donotkill_sysproc, is the flag to enable / disable this feature for
the
system processes chosen by the user
- donotkill_proc_names, is the list of process names (comma separated)
chosen by the user to be preserved from killing
- donotkill_sysproc_names, is the list of user system process names
(comma separated)
chosen by the user to be preserved from killing

When a the lowmemorykiller needs to free ram, it will first kill the
processes
not included inside the two process name lists chosen by the user.
If there are no more standard killable processes, then it will be
necessary to
kill one of the processes included inside one of the two white-lists, to
prevent system hangs, slowdowns, etc.

Change-Id: Ic1e10f4ebbfce22cab6566887a3101839afeff22